Lumen Health: the Clinical dashboard, built on a weekday heat grid
Fourth theme, and deliberately unlike the other three: white cards on a cool
grey, 16px body, generous spacing, nothing dense. A practice manager reads this
on a laptop between patients.

Signature visual is the consulting week as a grid of weekday against hour,
built as a real table so it keeps its headers, reads to a screen reader and
copies into a spreadsheet. Two grids, two hues, because slot fill and waiting
time are different quantities and sharing a ramp would invite reading them as
one. They come out near mirror images, which is the finding.

Palette validated with the dataviz script before any CSS was written.
Categorical #00919A #C0521C #6B46C6 #57802C clears delta E 25.4 normal and 8.3
tritan on white. Both six step ramps pass as ordinal: monotone lightness, every
adjacent gap over 0.06, and a light end clearing 2:1 against the surface, which
is why neither starts near white. Cell text switches to dark ink below the
fourth step, checked against the hexes rather than guessed.

Three things found by rendering it and looking:

  The patient collection chart was a blended rate that contradicted its own
  copy. Self funded charges collect at 89% and scheme gaps at 50%; averaging
  them gives 72% at every site, which describes neither. Added
  mart_patient_charges at site by charge type. The two clusters never cross, so
  the finding is a process rather than a site, and the page now says that.

  DECIMAL columns reach the browser from Arrow as an object, not a number, and
  the first symptom is toFixed throwing inside a formatter several layers away.
  Fixed at the model with an explicit cast, and db.js now converts any decimal
  at the boundary so the next one cannot get through.

  A figure had a table and no chart, leaving an empty card, and both reference
  line labels were drawn under a tall bar at the right edge. Labels moved into
  the legend.

159 tests pass, 2 deliberate warnings. The landing page now carries the real
headline number for all four demos instead of three placeholders.

+++ projects/04-lumen-health/dbt/models/gold/mart_capacity_cost.sql
@@ -86,7 +86,10 @@ final as (
         round(s.slots_no_show * s.cost_per_slot_zar, 2) as no_show_cost_zar,
         round(s.slots_excess * s.cost_per_slot_zar, 2) as excess_cost_zar,
 
-        {{ target }} * 100 as target_fill_pct,
+        -- Cast explicitly: Jinja renders this as decimal arithmetic, DuckDB types the
+        -- result DECIMAL, and Arrow hands a decimal to the browser as an object rather
+        -- than a number. Every numeric column that crosses into JavaScript is a double.
+        ({{ target }} * 100)::double as target_fill_pct,
         t.all_slots,
         t.all_booked,
         t.all_unfilled,
